Rewrote terminal configuration functions into seperate files and functions
