+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         ::::::::             #
 #    Makefile                                           :+:    :+:             #
 #                                                      +:+                     #
-#    By: omulder <omulder@student.codam.nl>           +#+                      #
+#    By: jbrinksm <jbrinksm@student.codam.nl>         +#+                      #
 #                                                    +#+                       #
 #    Created: 2019/04/10 20:30:07 by jbrinksm       #+#    #+#                 #
-#    Updated: 2019/04/18 17:33:37 by omulder       ########   odam.nl          #
+#    Updated: 2019/04/18 19:10:55 by jbrinksm      ########   odam.nl          #
 #                                                                              #
 # **************************************************************************** #
 
@@ -18,7 +18,9 @@ INCLUDES = -I./ -I./libft/ -I./includes -I../includes
 LIBFT= ./libft/libft.a
 LIB = -L./libft/ -lft -ltermcap
 VPATH = ./srcs ./test ./libft
-SRCS = term_init shell_prompt builtin_exit input_read parser_lexer jornfuckup
+SRCS = shell_prompt builtin_exit input_read parser_lexer jornfuckup \
+term_prepare term_is_valid term_init_struct term_get_attributes \
+term_set_attributes term_reset_attributes term_free_struct 
 TESTS = test_main.c
 OBJECTS := $(SRCS:%=%.o)
 TESTOBJECTS := $(TESTS:%.c=%.o)
