+++ srcs/term_free_struct.c
@@ -1,20 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   builtin_exit.c                                     :+:    :+:            */
+/*   term_free_struct.c                                 :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: jbrinksm <jbrinksm@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
-/*   Created: 2019/04/11 20:15:24 by jbrinksm       #+#    #+#                */
-/*   Updated: 2019/04/18 19:15:46 by jbrinksm      ########   odam.nl         */
+/*   Created: 2019/04/18 18:18:53 by jbrinksm       #+#    #+#                */
+/*   Updated: 2019/04/18 18:19:19 by jbrinksm      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vsh.h"
 
-void	builtin_exit(t_term *term_p)
+void	term_free_struct(t_term *term_p)
 {
-	ft_putchar('\n');
-	term_free_struct(term_p);
-	exit(FUNCT_SUCCESS);
+	if (term_p)
+	{
+		if (term_p->old_termios_p)
+			free(term_p->old_termios_p);
+		if (term_p->termios_p)
+			free(term_p->termios_p);
+		free(term_p);
+	}
 }
