+++ srcs/term_reset_attributes.c
@@ -1,20 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   builtin_exit.c                                     :+:    :+:            */
+/*   term_reset_attributes.c                            :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: jbrinksm <jbrinksm@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
-/*   Created: 2019/04/11 20:15:24 by jbrinksm       #+#    #+#                */
-/*   Updated: 2019/04/18 19:15:46 by jbrinksm      ########   odam.nl         */
+/*   Created: 2019/04/18 18:19:47 by jbrinksm       #+#    #+#                */
+/*   Updated: 2019/04/18 18:19:56 by jbrinksm      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vsh.h"
 
-void	builtin_exit(t_term *term_p)
+int		term_reset(t_term *term_p)
 {
-	ft_putchar('\n');
-	term_free_struct(term_p);
-	exit(FUNCT_SUCCESS);
+	int	ret;
+
+	/* Is TCSANOW flag correct? */
+	ret = tcsetattr(STDIN_FILENO, TCSANOW, term_p->old_termios_p);
+	if (ret == -1)
+		return (FUNCT_SUCCESS);
+	return (FUNCT_SUCCESS);
 }
